Simplify SetSource to be IsRootContext
