+++ src/Microsoft.AspNet.Hosting/IHttpContextAccessor.cs
@@ -1,14 +1,15 @@
 // Copyright (c) Microsoft Open Technologies, Inc. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using Microsoft.AspNet.Http;
-using Microsoft.Framework.DependencyInjection;
 
 namespace Microsoft.AspNet.Hosting
 {
-    public interface IHttpContextAccessor : IContextAccessor<HttpContext>
+    public interface IHttpContextAccessor
     {
-        IDisposable SetSource(Func<HttpContext> access, Func<HttpContext, HttpContext> exchange);
+        bool IsRootContext { get; set; }
+        HttpContext Value { get; }
+
+        HttpContext SetValue(HttpContext value);
     }
 }
