+++ src/Microsoft.AspNet.RequestContainer/RequestServicesContainer.cs
@@ -2,13 +2,9 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-#if ASPNET50
-using System.Runtime.Remoting.Messaging;
-using System.Runtime.Remoting;
-#endif
+using Microsoft.AspNet.Hosting;
 using Microsoft.AspNet.Http;
 using Microsoft.Framework.DependencyInjection;
-using Microsoft.AspNet.Hosting;
 
 namespace Microsoft.AspNet.RequestContainer
 {
@@ -59,27 +55,6 @@ namespace Microsoft.AspNet.RequestContainer
         private IHttpContextAccessor ScopeContextAccessor { get; set; }
         private IHttpContextAccessor AppContextAccessor { get; set; }
 
-        private const string LogicalDataKey = "__HttpContext_Current__";
-        private static HttpContext AccessRootHttpContext()
-        {
-#if ASPNET50
-            var handle = CallContext.LogicalGetData(LogicalDataKey) as ObjectHandle;
-            return handle != null ? handle.Unwrap() as HttpContext : null;
-#else
-            throw new Exception("TODO: CallContext not available");
-#endif 
-        }
-
-        private static HttpContext ExchangeRootHttpContext(HttpContext httpContext)
-        {
-#if ASPNET50
-            var prior = CallContext.LogicalGetData(LogicalDataKey) as ObjectHandle;
-            CallContext.LogicalSetData(LogicalDataKey, new ObjectHandle(httpContext));
-            return prior != null ? prior.Unwrap() as HttpContext : null;
-#else
-            return null;
-#endif
-        }
 
         // CONSIDER: this could be an extension method on HttpContext instead
         public static RequestServicesContainer EnsureRequestServices(HttpContext httpContext, IServiceProvider services)
@@ -101,7 +76,7 @@ namespace Microsoft.AspNet.RequestContainer
             var rootHttpContextAccessor = serviceProvider.GetRequiredService<IHttpContextAccessor>();
             var rootServiceScopeFactory = serviceProvider.GetRequiredService<IServiceScopeFactory>();
 
-            rootHttpContextAccessor.SetSource(AccessRootHttpContext, ExchangeRootHttpContext);
+            rootHttpContextAccessor.IsRootContext = true;
 
             // Pre Scope setup
             var priorApplicationServices = serviceProvider;
@@ -123,7 +98,7 @@ namespace Microsoft.AspNet.RequestContainer
             return new RequestServicesContainer(httpContext, appServiceScopeFactory, appHttpContextAccessor, appServiceProvider);
         }
 
-        #region IDisposable Support
+#region IDisposable Support
         private bool disposedValue = false; // To detect redundant calls
 
         protected virtual void Dispose(bool disposing)
@@ -163,6 +138,6 @@ namespace Microsoft.AspNet.RequestContainer
             // Do not change this code. Put cleanup code in Dispose(bool disposing) above.
             Dispose(true);
         }
-        #endregion
+#endregion
     }
 }
