+++ src/Microsoft.AspNet.RequestContainer/project.json
@@ -7,7 +7,15 @@
     "Microsoft.Framework.OptionsModel": "1.0.0-*"
   },
   "frameworks": {
-    "aspnet50": {},
-    "aspnetcore50": {}
+    "aspnet50": {
+      "dependencies": {
+        "System.Threading": "4.0.10-beta-*"
+      }
+    },
+    "aspnetcore50": {
+      "dependencies": {
+        "System.Threading": "4.0.10-beta-*"
+      }
+    },
   }
 }
